+++ packages/react/src/aspect-ratio/index.ts
@@ -1,0 +1,2 @@
+export { AspectRatio } from './AspectRatio.styled';
+export type { AspectRatioProps } from './AspectRatio.styled';