+++ packages/react/src/blockquote/index.ts
@@ -1,0 +1,7 @@
+export * as Blockquote from './Blockquote.styled';
+
+export type {
+  BlockquoteRootProps,
+  BlockquoteContentProps,
+  BlockquoteAttributionProps,
+} from './Blockquote.styled';