+++ packages/react/src/skeleton/index.ts
@@ -1,0 +1,3 @@
+export { Skeleton } from './Skeleton.styled';
+
+export type { SkeletonProps } from './Skeleton.styled';